move wps process validation to separate classes
